+++ index.js
@@ -13,14 +13,10 @@ module.exports = {
           baseIndent: 1,
           switchCase: 1,
         }],
-        // 防止使用未知的 DOM 属性
-        //  - 在 Vue 中使用 JSX 时, 支持使用 class 设置样式类
-        'react/no-unknown-property': ['error', {
-          ignore: ['class'],
-        }],
 
         // 仅在 React 中需要的规则, 在 Vue 中不需要
         'react/prop-types': 'off',
+        'react/no-unknown-property': 'off',
       },
     },
   ],